Desktop/resume_rag_chatbot/requirements.txt
streamlit==1.29.0
chromadb==0.4.22
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6
